split util
