+++ src/Spectre.AngularClient/src/app/divik-results/divik/divik.component.ts
@@ -17,12 +17,13 @@
    limitations under the License.
 */
 
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute} from '@angular/router';
-import { DivikService } from '../shared/divik.service';
-import { Heatmap } from '../../heatmaps/shared/heatmap';
-import { DivikConfig } from '../shared/divik-config';
 import { DomSanitizer } from '@angular/platform-browser';
+
+import { Heatmap } from '../../heatmaps/shared/heatmap';
+import { DivikService } from '../shared/divik.service';
+import { DivikConfig } from '../shared/divik-config';
 
 
 @Component({
