+++ src/Spectre.AngularClient/src/app/heatmaps/shared/heatmap.service.ts
@@ -17,71 +17,34 @@
  limitations under the License.
  */
 
-import { Injectable } from '@angular/core';
-import { Http, Response, Headers } from '@angular/http';
-import { Observable } from 'rxjs/Rx';
+import {Injectable} from '@angular/core';
+import {Http, Response, Headers} from '@angular/http';
+import {Observable} from 'rxjs/Rx';
 import 'rxjs/Rx';
 
-import { Heatmap } from './heatmap';
+import {Heatmap} from './heatmap';
 import {Service} from '../../app.service';
+import {HeatmapUtil} from './heatmap-util';
 
 @Injectable()
 export class HeatmapService extends Service {
 
-    constructor(private http: Http) { super(); }
+  constructor(private http: Http) {
+    super();
+  }
 
-    private getHeaders() {
-      const headers = new Headers();
-      headers.append('Accept', 'application/json');
-      return headers;
-    }
+  private getHeaders() {
+    const headers = new Headers();
+    headers.append('Accept', 'application/json');
+    return headers;
+  }
 
-    get(preparationId: number, channelId: number): Observable<Heatmap> {
-      console.log('[HeatmapService] preparationId: ' + preparationId + ' channelId ' + channelId);
-      const queryUrl = `${this.getBaseUrl()}/heatmap/${preparationId}?channelId=${channelId}&flag=false`;
-      const response = this.http.get(queryUrl, {headers: this.getHeaders()});
-      const spectrum = response.map((res: Response) => HeatmapUtil.toHeatmap(res));
-      return spectrum;
-    }
-}
-
-export class HeatmapUtil {
-  static toHeatmap(response: Response, className = '[HeatmapService]'): Heatmap {
-    console.log(className + ' parsing response:');
-    console.log(response);
-    const json = response.json();
-    console.log(json);
-    const max_column = Math.max.apply(null, json.X);
-    const min_column = Math.min.apply(null, json.X);
-    const max_row = Math.max.apply(null, json.Y);
-    const min_row = Math.min.apply(null, json.Y);
-    console.log(className + ' found bounds: (' + min_row + ', ' + max_row + '), (' + min_column + ', ' + max_column + ')');
-    console.log(className + ' initializing heatmap');
-    const data = [];
-    for (let i = min_row; i < max_row + 1; i++) {
-      data[i - min_row] = [];
-      for (let j = min_column; j < max_column + 1; j++) {
-        data[i - min_row][j - min_column] = 0;
-      }
-    }
-    console.log(className + ' processing heatmap');
-    for (let i = 0; i < json.X.length; i++) {
-      const column = json.X[i] - min_column;
-      const row = max_row - json.Y[i];
-      if (className === '[HeatmapService]') {
-        const intensity = json.Intensities[i];
-        console.log(className + ' processing (' + row + ', ' + column + ', ' + intensity + ')');
-        data[row][column] = intensity;
-      } else {
-        const jsonData = json.Data[i];
-        console.log(className + ' processing (' + row + ', ' + column + ', ' + jsonData + ')');
-        data[row][column] = jsonData;
-      }
-    }
-    console.log(className + ' heatmap processed');
-    return <Heatmap>({
-      data: data
-    });
+  get(preparationId: number, channelId: number): Observable<Heatmap> {
+    console.log('[HeatmapService] preparationId: ' + preparationId + ' channelId ' + channelId);
+    const queryUrl = `${this.getBaseUrl()}/heatmap/${preparationId}?channelId=${channelId}&flag=false`;
+    const response = this.http.get(queryUrl, {headers: this.getHeaders()});
+    const spectrum = response.map((res: Response) => HeatmapUtil.toHeatmap(res));
+    return spectrum;
   }
 }
